Refactor some functions to fix compile errors
